Rearrange the variant registration code for immediate registration
