+++ src/core/task/registrar.h
@@ -20,7 +20,7 @@
 
 #include "legion.h"
 
-#include "core/task/variant.h"
+#include "core/task/variant_options.h"
 #include "core/utilities/typedefs.h"
 
 /**
