+++ src/core/task/variant_helper.h
@@ -18,69 +18,10 @@
 
 #include "legion.h"
 
-#include "core/utilities/typedefs.h"
+#include "core/task/task_info.h"
+#include "core/task/variant_options.h"
 
-/**
- * @file
- * @brief Class definition fo legate::VariantOptions
- */
 namespace legate {
-
-/**
- * @brief Function signature for task variants. Each task variant must be a function of this type.
- */
-class TaskContext;
-using VariantImpl = void (*)(TaskContext&);
-
-// Each scalar output store can take up to 12 bytes, so in the worst case there can be only up to
-// 341 scalar output stores.
-constexpr size_t LEGATE_MAX_SIZE_SCALAR_RETURN = 4096;
-
-/**
- * @ingroup task
- * @brief A helper class for specifying variant options
- */
-struct VariantOptions {
-  /**
-   * @brief If the flag is `true`, the variant launches no subtasks. `true` by default.
-   */
-  bool leaf{true};
-  bool inner{false};
-  bool idempotent{false};
-  /**
-   * @brief If the flag is `true`, the variant needs a concurrent task launch. `false` by default.
-   */
-  bool concurrent{false};
-  /**
-   * @brief Maximum aggregate size for scalar output values. 4096 by default.
-   */
-  size_t return_size{LEGATE_MAX_SIZE_SCALAR_RETURN};
-
-  /**
-   * @brief Changes the value of the `leaf` flag
-   *
-   * @param `leaf` A new value for the `leaf` flag
-   */
-  VariantOptions& with_leaf(bool leaf);
-  VariantOptions& with_inner(bool inner);
-  VariantOptions& with_idempotent(bool idempotent);
-  /**
-   * @brief Changes the value of the `concurrent` flag
-   *
-   * @param `concurrent` A new value for the `concurrent` flag
-   */
-  VariantOptions& with_concurrent(bool concurrent);
-  /**
-   * @brief Sets a maximum aggregate size for scalar output values
-   *
-   * @param `return_size` A new maximum aggregate size for scalar output values
-   */
-  VariantOptions& with_return_size(size_t return_size);
-
-  void populate_registrar(Legion::TaskVariantRegistrar& registrar);
-};
-
-std::ostream& operator<<(std::ostream& os, const VariantOptions& options);
 
 namespace detail {
 
@@ -120,11 +61,19 @@ struct RegisterVariantImpl {
   {
     T::BASE::template register_variant<SELECTOR<T>::variant>(SELECTOR<T>::id, options);
   }
+  static void add_variant(TaskInfo* task_info, const VariantOptions& options)
+  {
+    T::BASE::template add_variant<SELECTOR<T>::variant>(task_info, SELECTOR<T>::id, options);
+  }
 };
 
 template <typename T, template <typename...> typename SELECTOR>
 struct RegisterVariantImpl<T, SELECTOR, false> {
   static void register_variant(const VariantOptions& options)
+  {
+    // Do nothing
+  }
+  static void add_variant(TaskInfo* task_info, const VariantOptions& options)
   {
     // Do nothing
   }
@@ -135,6 +84,10 @@ struct RegisterVariant {
   static void register_variant(const VariantOptions& options)
   {
     RegisterVariantImpl<T, SELECTOR, SELECTOR<T>::value>::register_variant(options);
+  }
+  static void add_variant(TaskInfo* task_info, const VariantOptions& options)
+  {
+    RegisterVariantImpl<T, SELECTOR, SELECTOR<T>::value>::add_variant(task_info, options);
   }
 };
